[Refactor] Move BitmapValue from package spark-dpp to plugin-common (#36533)
