+++ fe/plugin-common/src/main/java/com/starrocks/types/BitmapValue.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package com.starrocks.load.loadv2.dpp;
+package com.starrocks.types;
 
 import com.google.common.base.Objects;
 import org.roaringbitmap.Util;
