+++ fe/plugin-common/src/test/java/com/starrocks/types/BitmapValueTest.java
@@ -15,9 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package com.starrocks.load.loadv2.dpp;
+package com.starrocks.types;
 
-import com.starrocks.common.Codec;
 import org.junit.Assert;
 import org.junit.Test;
