+++ fe/spark-dpp/src/main/java/com/starrocks/load/loadv2/dpp/SparkRDDAggregator.java
@@ -19,6 +19,7 @@ package com.starrocks.load.loadv2.dpp;
 
 import com.starrocks.common.SparkDppException;
 import com.starrocks.load.loadv2.etl.EtlJobConfig;
+import com.starrocks.types.BitmapValue;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.log4j.LogManager;
